Fixed: Structure parser failing on files with Unix line ending
The line is reported as empty for files with Unix line ending (LF)
and this was treated as if this indicated the end of the file.
Checking the rdstate is the proper way of checking whether
we can still read from that buffer, or whether it is exhausted
or some other error occured.