The preview image tool window CPreviewImageView has been added. It shows the preview image and adjusts its size. It still needs more work such as a toolbar with a progress bar and options.
All user messages, starting with WM_USER or WM_APP, have been centralized.
This avoids conflicts and allows an overview over these messages.
They can now be found under the enum AfxUserMessages.

A new user message to the mainframe allows to show
any tool window, not just toggle it. OnShowDockingBar()
is the function that is called in response to that message.

The OutputBuilder/WorkerThread can now send several messages after termination,
not just one. Its member MsgsAfterTermination is responsible for that.

The symbol for the BookmarkView has been changed slightly.